requirements.txt
# List here the requirements for running this package.
django>=1.11
django-model-utils>=3.0

